% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bai_pinus.R
\docType{data}
\name{bai_pinus}
\alias{bai_pinus}
\title{Basal Area Index of Mediterranean \emph{Pinus} species}
\format{
A data frame with 564 rows and 5 columns:
\itemize{
\item \code{bai}: basal area index (\eqn{cm^2 \cdot year^{-1}}) of all trees of the studied plots
\item \code{sp_code}: code of the species studied
\item \code{prec}: annual precipitation (mm)
\item \code{tmed}: annual mean temperature (º)
\item \code{water_balance}: difference among the annual precipitation and the potential
evapotransporation.
}
}
\usage{
bai_pinus
}
\description{
A dataset with Basal Area Index (bai) of several \emph{Pinus} species plots and
climate variables for each year.
}
\keyword{datasets}
