% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodTools-package.R
\docType{package}
\name{likelihoodTools-package}
\alias{likelihoodTools}
\alias{likelihoodTools-package}
\title{likelihoodTools: Tools for managing results from maximum likelihood estimation}
\description{
Tools for managing and exploring parameter estimation results derived from Maximum Likelihood Estimation (MLE) using the `likelihood` package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ajpelu/likelihoodTools}
  \item \url{https://ajpelu.github.io/likelihoodTools/}
  \item Report bugs at \url{https://github.com/ajpelu/likelihoodTools/issues/}
}

}
\author{
\strong{Maintainer}: Antonio Jesús Pérez-Luque \email{ajpelu@gmail.com} (\href{https://orcid.org/0000-0002-1747-0469}{ORCID}) [copyright holder]

}
\keyword{internal}
